requirements.txt
Flask
Werkzeug
Flask-WTF
requests
openai
gunicorn
Flask-Cors
numpy
pydub
SoundFile
# opensoundscape # Temporarily removed again to debug load failure
# Removed version pins to allow uv to resolve dependencies
# If you need specific versions, consider installing them in a separate environment
# opensoundscape removed due to Python 3.12 incompatibility (via pathtools)
# RAG system dependencies
langchain>=0.0.235
chromadb>=0.3.25
sentence-transformers>=2.2.2
unstructured>=0.6.0
# Required by UnstructuredMarkdownLoader
markdown
google-generativeai
python-dotenv
langchain-huggingface # For HuggingFaceEmbeddings
langchain-chroma # For Chroma vector store integration
langchain-community # For document loaders like DirectoryLoader